Change rating calculator response to int

+++ src/Shared/SocialMedia/Events/PostSocialMediaResearched.php
@@ -13,7 +13,7 @@ class PostSocialMediaResearched
         public string $title,
         public string $body,
         public int $userId,
-        public float $rating,
+        public int $rating,
     ) {
     }
 }
